Refactor profiling request pydantic models to its correct file
