+++ src/gprofiler/backend/routers/metrics_routes.py
@@ -18,7 +18,7 @@ import math
 import uuid
 from datetime import datetime, timedelta
 from logging import getLogger
-from typing import List, Optional, Dict, Any, Literal
+from typing import List, Optional
 
 from botocore.exceptions import ClientError
 
@@ -32,6 +32,12 @@ from backend.models.metrics_models import (
     MetricNodesAndCores,
     MetricNodesCoresSummary,
     MetricSummary,
+    ProfilingResponse,
+    ProfilingRequest,
+    HeartbeatRequest,
+    HeartbeatResponse,
+    CommandCompletionRequest,
+    
     SampleCount,
     HTMLMetadata,
 )
@@ -39,92 +45,12 @@ from backend.utils.filters_utils import get_rql_first_eq_key, get_rql_only_for_o
 from backend.utils.request_utils import flamegraph_base_request_params, get_metrics_response, get_query_response
 from fastapi import APIRouter, Depends, Query, HTTPException
 from fastapi.responses import Response
-from pydantic import BaseModel, model_validator
 
 from gprofiler_dev import S3ProfileDal
 from gprofiler_dev.postgres.db_manager import DBManager
 
 logger = getLogger(__name__)
 router = APIRouter()
-
-
-class ProfilingRequest(BaseModel):
-    """Model for profiling request parameters"""
-    service_name: str
-    request_type: str = "start"  # start, stop
-    duration: Optional[int] = 60  # seconds
-    frequency: Optional[int] = 11  # Hz
-    profiling_mode: Optional[Literal["cpu", "allocation", "none"]] = "cpu"  # cpu, allocation, none
-    target_hostnames: Optional[List[str]] = None
-    pids: Optional[List[int]] = None
-    stop_level: Optional[str] = "process"  # process, host (only relevant for stop commands)
-    additional_args: Optional[Dict[str, Any]] = None
-
-    @model_validator(mode='after')
-    def validate_pids_for_process_stop(cls, model):
-        """Validate that PIDs are provided when command_type is stop and stop_level is process"""
-        if model.command_type == 'stop' and model.stop_level == 'process':
-            if not model.pids or len(model.pids) == 0:
-                raise ValueError('At least one PID must be provided when command_type is "stop" and stop_level is "process"')
-
-        # Validate the duration
-        if model.duration and model.duration <= 0:
-            raise ValueError("Duration must be a positive integer (seconds)")
-
-        # Validate the frequency
-        if model.frequency and model.frequency <= 0:
-            raise ValueError("Frequency must be a positive integer (Hz)")
-
-        return model
-
-
-class ProfilingResponse(BaseModel):
-    """Response model for profiling requests"""
-    success: bool
-    message: str
-    request_id: Optional[str] = None
-    command_ids: Optional[List[str]] = None  # List of command IDs for agent idempotency
-    estimated_completion_time: Optional[datetime] = None
-
-
-class HeartbeatRequest(BaseModel):
-    """Model for host heartbeat request"""
-    ip_address: str
-    hostname: str
-    service_name: str
-    last_command_id: Optional[str] = None
-    status: Optional[str] = "active"  # active, idle, error
-    timestamp: Optional[datetime] = None
-
-
-class HeartbeatResponse(BaseModel):
-    """Response model for heartbeat requests"""
-    success: bool
-    message: str
-    profiling_command: Optional[Dict[str, Any]] = None  # Changed from profiling_request to profiling_command
-    command_id: Optional[str] = None
-
-
-class ProfilingCommand(BaseModel):
-    """Model for combined profiling command sent to hosts"""
-    command_id: str
-    command_type: str  # start, stop
-    hostname: str
-    service_name: str
-    combined_config: Dict[str, Any]  # Combined configuration from multiple requests
-    request_ids: List[str]  # List of request IDs that make up this command
-    created_at: datetime
-    status: str  # pending, sent, completed, failed
-
-
-class CommandCompletionRequest(BaseModel):
-    """Model for reporting command completion"""
-    command_id: str
-    hostname: str
-    status: Literal["completed", "failed"]  # completed, failed
-    execution_time: Optional[int] = None  # seconds
-    error_message: Optional[str] = None
-    results_path: Optional[str] = None  # S3 path or local path to results
 
 
 def get_time_interval_value(start_time: datetime, end_time: datetime, interval: str) -> str:
@@ -319,7 +245,7 @@ def get_html_metadata(
 
 
 @router.post("/profile_request", response_model=ProfilingResponse)
-def create_profiling_request(profiling_request: ProfilingRequest):
+def create_profiling_request(profiling_request: ProfilingRequest) -> ProfilingResponse:
     """
     Create a new profiling request with the specified parameters.
